Move Graphics settings to `menus_settings_graphics.cpp`
