+++ CMakeLists.txt
@@ -2544,6 +2544,7 @@ if(CLIENT)
     components/menus_settings_assets.cpp
     components/menus_settings_controls.cpp
     components/menus_settings_controls.h
+    components/menus_settings_graphics.cpp
     components/menus_settings_player.cpp
     components/menus_settings_tee.cpp
     components/menus_settings_tee7.cpp
